+++ apps/website/src/module/root/ui/page/index.tsx
@@ -1,0 +1,1 @@
+export { RootPage } from './page.presenter';
